Render thread

+++ src/primitives/line/line.hpp
@@ -43,7 +43,8 @@ namespace bolt
     {
         glBindBuffer(GL_ARRAY_BUFFER, buffer);
 
-        std::vector<float> data = {p1.x, p1.y, p1.z, p2.x, p2.y, p2.z};
+        std::vector<float> data = {static_cast<float>(p1.x), static_cast<float>(p1.y), static_cast<float>(p1.z), 
+                                    static_cast<float>(p2.x), static_cast<float>(p2.y), static_cast<float>(p2.z)};
 
         glBufferSubData(GL_ARRAY_BUFFER, current_size, 6 * sizeof(float), data.data());
 
@@ -65,7 +66,8 @@ namespace bolt
 
         constexpr std::size_t arg_num = sizeof...(Args); // number of args
 
-        std::vector<float> data = {p1.x, p1.y, p1.z, p2.x, p2.y, p2.z};
+        std::vector<float> data = {static_cast<float>(p1.x), static_cast<float>(p1.y), static_cast<float>(p1.z), 
+                                    static_cast<float>(p2.x), static_cast<float>(p2.y), static_cast<float>(p2.z)};
 
         glBufferData(GL_ARRAY_BUFFER, 6 * sizeof(float) + arg_num * 3 * sizeof(float), data.data(), GL_STATIC_DRAW);
